Use partial class files instead of ifdefs
